OlgaTPark/tenfourfox#12 — M1388020 (adapted patch)
This commit "tries" to mitigate M1388020 if ever present in TenFourFox (at least TenSixFox is using IPC gfx for plugins, so, this justifies the patch).  The patch has been adapted to do what it's possible to do here and could be partial because the code changed between Firefox 45 and Firefox 52.

References:
https://bugzilla.mozilla.org/show_bug.cgi?id=1388020
https://hg.mozilla.org/releases/mozilla-esr52/rev/e8ab2736499b
https://hg.mozilla.org/releases/mozilla-esr52/rev/ad9a885b0df4